+++ tests/testthat/test-algebraNA-alg_properties.R
@@ -31,12 +31,12 @@ test_that('algebraic properties', {
                     expect_equal(a$i(x, y), a$i(y, x), info=amodif)
                     expect_equal(a$pi(x, y), a$pi(y, x), info=amodif)
 
+                    # bi-residuum
+                    expect_equal(a$b(x, y), a$i(a$r(x, y), a$r(y, x)), info=amodif)
+
                     if (!(amodif %in% c('dragonfly', 'lowerEst'))) {
                         # De-Morgan's law
                         expect_equal(a$c(!!!x, !!!y), a$ni(a$t(a$ni(!!!x), a$ni(!!!y))), info=amodif)
-
-                        # bi-residuum
-                        expect_equal(a$b(x, y), a$i(a$r(x, y), a$r(y, x)), info=amodif)
                     }
                 }
             }